fix setup form add dark mode and arrange files
