+++ public/src/app.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Dashboard from "./components/dashboard";
+import Dashboard from "./components/pages/dashboard";
 
 ReactDOM.render(<Dashboard />, document.getElementById("root"));
