+++ public/src/components/page_objects/simpleLineChart.js
