+++ public/src/components/page_objects/simpleTable.js
@@ -1,10 +1,10 @@
-import Paper from "@material-ui/core/Paper";
-import {withStyles} from "@material-ui/core/styles";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
+import Paper from "@material-ui/core/Paper/index";
+import { withStyles } from "@material-ui/core/styles/index";
+import Table from "@material-ui/core/Table/index";
+import TableBody from "@material-ui/core/TableBody/index";
+import TableCell from "@material-ui/core/TableCell/index";
+import TableHead from "@material-ui/core/TableHead/index";
+import TableRow from "@material-ui/core/TableRow/index";
 import PropTypes from "prop-types";
 import React from "react";
