+++ public/src/components/pages/contentItemView.js
