+++ public/src/components/pages/pollSetup.js
@@ -1,5 +1,5 @@
-import Button from "@material-ui/core/Button";
-import TextField from "@material-ui/core/TextField";
+import Button from "@material-ui/core/Button/index";
+import TextField from "@material-ui/core/TextField/index";
 import React from "react";
 
 class PollSetup extends React.Component {
