+++ public/src/components/response.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {FormControl, FormGroup, FormLabel} from "react-bootstrap";
+import { FormControl, FormGroup, FormLabel } from "react-bootstrap";
 
 const EmptyResponse = (
   <div>
